moving all ping_pong related code into its own file
